Allow non-transform fixtures

+++ test/helpers/fixtures-runner.js
@@ -1,4 +1,6 @@
 const path = require("path");
+const assert = require("assert");
+
 const {
   mapAsync,
   regG,
@@ -50,22 +52,36 @@ async function execFixture(pathname, loaderPath, useLoader) {
 }
 
 module.exports = async function runFixture(fixture, loaderPath) {
-  const compiledFolder = path.join(fixture.path, fixture.version);
   const srcFolder = path.join(fixture.path, "src");
 
-  test(`${fixture.name} - babel ${fixture.version}`, async () => {
-    await rimraf(compiledFolder);
-    await compileFiles(fixture);
+  if (fixture.transform) {
+    assert(fixture.version);
 
-    const actual = await execFixture(compiledFolder, loaderPath, true);
+    const compiledFolder = path.join(fixture.path, fixture.version);
 
-    if (fixture.type === "snapshot") {
+    test(`${fixture.name} - babel ${fixture.version}`, async () => {
+      await rimraf(compiledFolder);
+      await compileFiles(fixture);
+
+      const actual = await execFixture(compiledFolder, loaderPath, true);
+
+      if (fixture.type === "snapshot") {
+        expect(actual.stdout).toMatchSnapshot("stdout");
+        expect(actual.stderr).toMatchSnapshot("stderr");
+      } else {
+        const expected = await execFixture(srcFolder, loaderPath);
+        expect(actual.stderr).toBe(expected.stderr);
+        expect(actual.stdout).toBe(expected.stdout);
+      }
+    });
+  } else {
+    assert.strictEqual(fixture.type, "snapshot");
+
+    test(fixture.name, async () => {
+      const actual = await execFixture(srcFolder, loaderPath, true);
+
       expect(actual.stdout).toMatchSnapshot("stdout");
       expect(actual.stderr).toMatchSnapshot("stderr");
-    } else {
-      const expected = await execFixture(srcFolder, loaderPath);
-      expect(actual.stderr).toBe(expected.stderr);
-      expect(actual.stdout).toBe(expected.stdout);
-    }
-  });
+    });
+  }
 };
